Include parent package moody for existing packages
