--- src/main/java/commands/ExitCommand.java
@@ -1,19 +1,0 @@
-package commands;
-
-import commands.Command;
-import storage.Storage;
-import tasks.TaskList;
-import ui.Ui;
-
-public class ExitCommand extends Command {
-
-    @Override
-    public void execute(TaskList tasks, Ui ui, Storage storage) {
-        ui.showGoodbye();
-    }
-
-    @Override
-    public boolean isExit() {
-        return true;
-    }
-}
