--- src/main/java/commands/ListCommand.java
@@ -1,19 +1,0 @@
-package commands;
-
-import commands.Command;
-import storage.Storage;
-import tasks.TaskList;
-import ui.Ui;
-
-public class ListCommand extends Command {
-
-    @Override
-    public void execute(TaskList tasks, Ui ui, Storage storage) {
-        ui.showTaskList(tasks);
-    }
-
-    @Override
-    public boolean isExit() {
-        return false;
-    }
-}
