--- src/main/java/exceptions/InvalidCommandException.java
@@ -1,8 +1,0 @@
-package exceptions;
-
-public class InvalidCommandException extends Exception {
-
-    public InvalidCommandException(String message) {
-        super(message);
-    }
-}
