--- src/main/java/exceptions/TaskInputException.java
@@ -1,8 +1,0 @@
-package exceptions;
-
-public class TaskInputException  extends Exception {
-
-    public TaskInputException(String message) {
-        super(message);
-    }
-}
