--- src/main/java/exceptions/TaskOutOfBoundsException.java
@@ -1,8 +1,0 @@
-package exceptions;
-
-public class TaskOutOfBoundsException extends Exception {
-
-    public TaskOutOfBoundsException(String message) {
-        super(message);
-    }
-}
