+++ src/main/java/moody/commands/ExitCommand.java
@@ -1,0 +1,18 @@
+package moody.commands;
+
+import moody.storage.Storage;
+import moody.tasks.TaskList;
+import moody.ui.Ui;
+
+public class ExitCommand extends Command {
+
+    @Override
+    public void execute(TaskList tasks, Ui ui, Storage storage) {
+        ui.showGoodbye();
+    }
+
+    @Override
+    public boolean isExit() {
+        return true;
+    }
+}
