+++ src/main/java/moody/commands/ListCommand.java
@@ -1,0 +1,18 @@
+package moody.commands;
+
+import moody.storage.Storage;
+import moody.tasks.TaskList;
+import moody.ui.Ui;
+
+public class ListCommand extends Command {
+
+    @Override
+    public void execute(TaskList tasks, Ui ui, Storage storage) {
+        ui.showTaskList(tasks);
+    }
+
+    @Override
+    public boolean isExit() {
+        return false;
+    }
+}
