+++ src/main/java/moody/exceptions/InvalidCommandException.java
@@ -1,0 +1,8 @@
+package moody.exceptions;
+
+public class InvalidCommandException extends Exception {
+
+    public InvalidCommandException(String message) {
+        super(message);
+    }
+}
