+++ src/main/java/moody/exceptions/TaskInputException.java
@@ -1,0 +1,8 @@
+package moody.exceptions;
+
+public class TaskInputException  extends Exception {
+
+    public TaskInputException(String message) {
+        super(message);
+    }
+}
