+++ src/main/java/moody/exceptions/TaskOutOfBoundsException.java
@@ -1,0 +1,8 @@
+package moody.exceptions;
+
+public class TaskOutOfBoundsException extends Exception {
+
+    public TaskOutOfBoundsException(String message) {
+        super(message);
+    }
+}
