+++ src/main/java/moody/tasks/Deadline.java
@@ -1,0 +1,26 @@
+package moody.tasks;
+
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
+
+public class Deadline extends Task {
+    private static final DateTimeFormatter INPUT_FORMATTER = DateTimeFormatter.ofPattern("yyyy-MM-dd HHmm");
+    private static final DateTimeFormatter OUTPUT_FORMATTER = DateTimeFormatter.ofPattern("MMM dd yyyy h:mm a");
+    protected LocalDateTime by;
+
+    public Deadline(String description, LocalDateTime by) {
+        super(description);
+        this.by = by;
+    }
+
+    @Override
+    public String toFileFormat() {
+        return "D | " + super.toFileFormat() + " | " + this.by.format(INPUT_FORMATTER);
+    }
+
+    @Override
+    public String toString() {
+        String timeframe = String.format("(by: %s)", this.by.format(OUTPUT_FORMATTER));
+        return String.format("[D]%s %s", super.toString(), timeframe);
+    }
+}
