+++ src/main/java/moody/tasks/Todo.java
@@ -1,0 +1,18 @@
+package moody.tasks;
+
+public class Todo extends Task {
+
+    public Todo(String description) {
+        super(description);
+    }
+
+    @Override
+    public String toFileFormat() {
+        return "T | " + super.toFileFormat();
+    }
+
+    @Override
+    public String toString() {
+        return String.format("[T]%s", super.toString());
+    }
+}
