--- src/main/java/tasks/Task.java
@@ -1,32 +1,0 @@
-package tasks;
-
-public class Task {
-    protected String description;
-    protected boolean isDone;
-
-    public Task(String description) {
-        this.description = description;
-        this.isDone = false;
-    }
-
-    public String getStatusIcon() {
-        return (isDone ? "X" : " "); // mark done task with X
-    }
-
-    public void markAsDone() {
-        this.isDone = true;
-    }
-
-    public void markAsNotDone() {
-        this.isDone = false;
-    }
-
-    public String toFileFormat() {
-        return (isDone ? "1" : "0") + " | " + this.description;
-    }
-
-    @Override
-    public String toString() {
-        return String.format("[%s] %s", getStatusIcon(), this.description);
-    }
-}
