--- src/main/java/tasks/TaskList.java
@@ -1,35 +1,0 @@
-package tasks;
-
-import java.util.ArrayList;
-
-public class TaskList {
-    protected ArrayList<Task> tasks;
-
-    public TaskList() {
-        this.tasks = new ArrayList<>();
-    }
-
-    public TaskList(ArrayList<Task> tasks) {
-        this.tasks = tasks;
-    }
-
-    public ArrayList<Task> toArrayList() {
-        return this.tasks;
-    }
-
-    public void add(Task task) {
-        tasks.add(task);
-    }
-
-    public Task remove(int index) {
-        return tasks.remove(index);
-    }
-
-    public Task get(int index) {
-        return tasks.get(index);
-    }
-
-    public int size() {
-        return tasks.size();
-    }
-}
