--- src/main/java/tasks/Todo.java
@@ -1,20 +1,0 @@
-package tasks;
-
-import tasks.Task;
-
-public class Todo extends Task {
-
-    public Todo(String description) {
-        super(description);
-    }
-
-    @Override
-    public String toFileFormat() {
-        return "T | " + super.toFileFormat();
-    }
-
-    @Override
-    public String toString() {
-        return String.format("[T]%s", super.toString());
-    }
-}
